requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Django==5.0.2
django-anymail==10.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyarrow==15.0.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
requests==2.31.0
sendgrid==6.11.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
waitress==3.0.0
whitenoise==6.6.0
yarl==1.9.4
